+++ application/common/addresses.py
@@ -249,3 +249,13 @@ def create(
 
     # Finished
     return address
+
+
+def get_addresses_count():
+    """
+    Returns the quantity of addresses in the database
+    """
+
+    return Address.query(keys_only = True, batch_size = 100000).count()
+
+